requirements.txt
google-generativeai==0.8.5
pyttsx3==2.90
SpeechRecognition==3.14.2
PyAudio==0.2.14
Flask==3.1.0
pyautogui==0.9.54
pytube==15.0.0
gunicorn==21.2.0 